refactor: extract transaction settlement helpers
